// src/components/Chip/types.ts
export interface ChipProps {
  label: string;
  variant?: ChipVariant;
}

export type ChipVariant =
  | "normal"
  | "fighting"
  | "flying"
  | "poison"
  | "ground"
  | "rock"
  | "bug"
  | "ghost"
  | "steel"
  | "fire"
  | "water"
  | "grass"
  | "electric"
  | "psychic"
  | "ice"
  | "dragon"
  | "dark"
  | "fairy"
  | "stellar"
  | "unknown"
  | "shadow";